[BUGFIX] Support TYPO3 v10 HTTP handler stack
An empty 'handler' array in TYPO3_CONF_VARS.HTTP was introduced in
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61407
When an empty 'handler' array is passed to the guzzle client,
an exception is thrown as guzzle expects 'handler'
to be callable or unset.

  [ERROR] request="…"
  component="SFC.Staticfilecache.Service.ClientService":
  Problems in single request running: handler must be a callable / …/typo3_src-10.4.0/vendor/guzzlehttp/guzzle/src/Client.php:67

TYPO3 core initializes an handler stack from the handler
array. For compatibility reasons the staticfilecache
should do so as well.

Fixes: #251