Force proper cache file MIME Type via .htaccess
This adds a ForceType directive to the .htaccess template, to make sure
that Apache applies the correct cache headers configured by ModExpires.

In addition this also makes sure that Apache sends the proper
Content-Type header, if ModHeaders is not available or the Content-Type
header is not whitelisted for inclusion in the .htaccess.